logger to scheduler
